Misc issues due to re-architecturing #1 and additional fixes
The first and third items are handled.
The second one is not handled yet (I worked on this bug previously but
was not able to fix it)

Additional fixes:
Decreased z-index of cy-panzoom. In this way users are not able to use
panzoom when a modal is open.
Removed some unused selectors from sbgnviz.css